requirements.txt
rel==0.4.7
websocket_client==1.4.2
